refactor: streamline agent search utilities
