+++ apps/agent/src/utils/exa/index.ts
@@ -1,0 +1,25 @@
+export {
+  EXA_CATEGORIES,
+  EXA_LIVECRAWL_MODES,
+  EXA_SEARCH_TYPES,
+  DEFAULT_TIMEOUT_MS,
+} from "./constants.js";
+export {
+  exaSearchArgsSchema,
+  exaSearchResponseSchema,
+  type ExaSearchArgs,
+  type ExaSearchParsedArgs,
+  type ExaSearchResponse,
+  type ExaSearchResult,
+  type ExaSearchSubpage,
+} from "./schemas.js";
+export {
+  type ExaSearchNormalizedHighlight,
+  type ExaSearchNormalizedResult,
+  type ExaSearchNormalizedSubpage,
+  type ExaSearchToolError,
+  type ExaSearchToolResult,
+  type ExaSearchToolSuccess,
+  type PerformExaSearchOptions,
+} from "./types.js";
+export { performExaSearch } from "./search.js";
